Lots of changes

+++ SchemaCompiler.hs
@@ -23,7 +23,11 @@ compile = do
 
 writeTable :: (FilePath, Text) -> IO ()
 writeTable (path, content) = do
-    writeFile (cs path) (cs content)
+    alreadyExists <- Directory.doesFileExist path
+    existingContent <- if alreadyExists then readFile path else return ""
+    when (existingContent /= cs content) $ do
+        putStrLn $ "Updating " <> cs path
+        writeFile (cs path) (cs content)
 
 writeStub :: (FilePath, Text) -> IO ()
 writeStub (path, content) = do
@@ -35,7 +39,7 @@ section = "\n"
 compileTable table@(Table name attributes) =
     "-- This file is auto generated and will be overriden regulary. Please edit `src/Model/" <> tableNameToModelName name <> ".hs` to customize the Model"
     <> section
-    <> "{-# LANGUAGE TypeSynonymInstances, FlexibleInstances #-}"
+    <> "{-# LANGUAGE TypeSynonymInstances, FlexibleInstances, InstanceSigs, MultiParamTypeClasses, TypeFamilies #-}"
     <> section
     <> "module Model.Generated." <> tableNameToModelName name <> " where\n\n"
     <> "import Foundation.HaskellSupport\n"
@@ -43,24 +47,41 @@ compileTable table@(Table name attributes) =
     <> "import ClassyPrelude hiding (id) \n"
     <> "import Database.PostgreSQL.Simple\n"
     <> "import Database.PostgreSQL.Simple.FromRow\n"
-    <> section
-    <> compileTypeAlias table
+    <> "import Foundation.ControllerSupport (ParamName (..))\n"
+    <> "import qualified Data.Function\n"
     <> section
     <> compileGenericDataDefinition table
+    <> compileTypeAlias table
+    <> compileNewTypeAlias table
     <> section
     <> compileFromRowInstance table
     <> section
     <> section
     <> compileCreate table
     <> section
+    <> compileUpdate table
+    <> section
     <> compileFindAll table
+    <> section
+    <> compileFindOrNothing table
     <> section
     <> compileFind table
     <> section
     <> compileUnit table
     <> section
+    <> compileBuild table
+    <> section
     <> compileFindByAttributes table
     <> section
+    <> compileAttributeNames table
+    <> section
+    <> compileAssign table
+    <> section
+    <> compileIdentity table
+    <> section
+    <> compileCombine table
+    <> section
+    <> compileHasId table
 
 
 compileStub table@(Table name attributes) =
@@ -82,8 +103,7 @@ getStubFilePath :: Table -> FilePath
 getStubFilePath (Table name attributes) = "src/Model/" <> (cs $ tableNameToModelName name) <> ".hs"
 
 compileDataDefinition :: Table -> Text
-compileDataDefinition table@(Table name attributes) =
-		"data " <> tableNameToModelName name <> " = " <> tableNameToModelName name <> " { " <> compileFields attributes <> " }\n"
+compileDataDefinition table@(Table name attributes) = "data " <> tableNameToModelName name <> " = " <> tableNameToModelName name <> " { " <> compileFields attributes <> " }\n"
     where
 		compileFields :: [Attribute] -> Text
 		compileFields attributes = intercalate ", " $ map compileField attributes
@@ -104,6 +124,20 @@ compileTypeAlias table@(Table name attributes) =
 		compileField (Field fieldName fieldType) = haskellType fieldType
 		haskellType :: FieldType -> Text
 		haskellType (SerialField) = "Int"
+		haskellType (TextField _) = "Text"
+		haskellType (IntField) = "Int"
+
+
+compileNewTypeAlias :: Table -> Text
+compileNewTypeAlias table@(Table name attributes) =
+		"type New" <> tableNameToModelName name <> " = " <> tableNameToModelName name <> "' " <> compileFields attributes <> "\n"
+    where
+		compileFields :: [Attribute] -> Text
+		compileFields attributes = intercalate " " $ map compileField attributes
+		compileField :: Attribute -> Text
+		compileField (Field fieldName fieldType) = haskellType fieldType
+		haskellType :: FieldType -> Text
+		haskellType (SerialField) = "()"
 		haskellType (TextField _) = "Text"
 		haskellType (IntField) = "Int"
 
@@ -137,10 +171,49 @@ compileCreate table@(Table name attributes) =
                 SerialField -> Nothing
                 otherwise   -> Just $ fieldName <> " model"
     in
-        "create :: (?modelContext :: ModelContext) => " <> modelName <> " -> IO " <> modelName <> " \n"
-        <> "create model = do\n"
+        "instance CanCreate New" <> modelName <> " where\n"
+        <> indent (
+            "create :: (?modelContext :: ModelContext) => New" <> modelName <> " -> IO " <> modelName <> "\n"
+                <> "type Created New" <> modelName <> " = " <> modelName <> "\n"
+                <> "create model = do\n"
+                <> indent ("let (ModelContext conn) = ?modelContext\n"
+                    <> "result <- Database.PostgreSQL.Simple.query conn \"INSERT INTO " <> name <> " (" <> columns <> ") VALUES (" <> values <> ") RETURNING *\" (" <> bindings <> ")\n"
+                    <> "return (unsafeHead result)\n"
+                    )
+            )
+
+compileUpdate table@(Table name attributes) =
+    let
+        modelName = tableNameToModelName name
+        columns = intercalate ", " $ map toColumn attributes
+        toColumn (Field fieldName fieldType) = fieldName
+        values = intercalate ", " $ map toValue attributes
+        toValue (Field fieldName fieldType) =
+            case fieldType of
+                SerialField -> "DEFAULT"
+                otherwise   -> "?"
+        bindings :: Text
+        bindings =
+            let
+                bindingValues = (map fromJust $ filter isJust (map toBinding attributes)) <> (["id model"])
+            in
+                if (ClassyPrelude.length bindingValues == 1)
+                    then "Only (" <> (unsafeHead bindingValues) <> ")"
+                    else intercalate ", " bindingValues
+        toBinding (Field fieldName fieldType) =
+            case fieldType of
+                SerialField -> Nothing
+                otherwise   -> Just $ fieldName <> " model"
+        updates = intercalate ", " (map fromJust $ filter isJust $ map update attributes)
+        update (Field fieldName fieldType) =
+            case fieldType of
+                SerialField -> Nothing
+                otherwise -> Just $ fieldName <> " = ?"
+    in
+        "update :: (?modelContext :: ModelContext) => " <> modelName <> " -> IO " <> modelName <> " \n"
+        <> "update model = do\n"
         <> indent ("let (ModelContext conn) = ?modelContext\n"
-            <> "result <- Database.PostgreSQL.Simple.query conn \"INSERT INTO " <> name <> " (" <> columns <> ") VALUES (" <> values <> ") RETURNING *\" (" <> bindings <> ")\n"
+            <> "result <- Database.PostgreSQL.Simple.query conn \"UPDATE " <> name <> " SET " <> updates <> " WHERE id = ? RETURNING *\" (" <> bindings <> ")\n"
             <> "return (unsafeHead result)\n"
         )
 
@@ -157,18 +230,28 @@ compileFindAll table@(Table name attributes) =
         )
 
 
-compileFind table@(Table name attributes) =
+compileFindOrNothing table@(Table name attributes) =
     let
         modelName = tableNameToModelName name
     in
-        "find :: (?modelContext :: ModelContext) => Int -> IO (Maybe " <> modelName <> ")\n"
-        <> "find id = do\n"
+        "findOrNothing :: (?modelContext :: ModelContext) => Int -> IO (Maybe " <> modelName <> ")\n"
+        <> "findOrNothing id = do\n"
         <> indent (
             "let (ModelContext conn) = ?modelContext\n"
             <> "results <- Database.PostgreSQL.Simple.query conn \"SELECT * FROM " <> name <> " WHERE id = ?\" [id]\n"
             <> "return $ headMay results\n"
         )
 
+compileFind table@(Table name attributes) =
+    let
+        modelName = tableNameToModelName name
+    in
+        "find :: (?modelContext :: ModelContext) => Int -> IO " <> modelName <> "\n"
+        <> "find id = do\n"
+        <> indent (
+            "result <- findOrNothing id\n"
+            <> "return (fromMaybe (error \"Model cannot be found \") result)"
+        )
 
 compileFromRowInstance table@(Table name attributes) =
     "instance FromRow " <> tableNameToModelName name <> " where \n"
@@ -183,6 +266,20 @@ compileUnit table@(Table name attributes) =
 		compileFields attributes = intercalate " " $ map compileField attributes
 		compileField :: Attribute -> Text
 		compileField _ = "()"
+
+compileBuild _ = "build = unit\n"
+
+compileIdentity :: Table -> Text
+compileIdentity table@(Table name attributes) =
+		"identity = " <> tableNameToModelName name <> " " <> compileFields attributes <> "\n"
+    where
+		compileFields :: [Attribute] -> Text
+		compileFields attributes = intercalate " " $ map compileField attributes
+		compileField :: Attribute -> Text
+		compileField _ = "Data.Function.id"
+
+compileHasId :: Table -> Text
+compileHasId table@(Table name attributes) = "instance HasId " <> tableNameToModelName name <> " where getId model = id model\n"
 
 compileFindByAttributes table@(Table tableName attributes) =
         intercalate "\n\n" $ map compileFindByAttribute attributes
@@ -202,8 +299,37 @@ compileFindByAttributes table@(Table tableName attributes) =
 
         compileFindByAttribute _ = ""
 
+compileAttributeNames table@(Table tableName attributes) =
+        "data Field = " <> (intercalate " | " (map compileAttributeName attributes)) <> " deriving (Show)"
+        <> section
+        <> section
+        <> "instance ParamName Field where \n" <> (intercalate "\n" (map compileParamName attributes)) <> "\n"
+        <> "instance FormField Field where \n"
+        <> "    type Model Field = " <> tableNameToModelName tableName <> "\n"
+        <> (intercalate "\n" (map compileFormFieldName attributes))
+        <> section
+        <> (intercalate "\n" (map compileFormFieldValue attributes))
+    where
+        compileAttributeName (Field name _) = tableNameToModelName name
+        compileParamName (Field name _) = indent $ "paramName " <> (tableNameToModelName name) <> " = \"" <> name <> "\""
+        compileFormFieldName (Field name _) = indent $ "formFieldName " <> (tableNameToModelName name) <> " = \"" <> name <> "\""
+        compileFormFieldValue (Field name _) = indent $ "formFieldValue " <> (tableNameToModelName name) <> " (" <> tableNameToModelName tableName <> " { " <> name <> " }) = inputValue " <> name
 
+compileAssign table@(Table tableName attributes) =
+        ""
+        --"assignField :: Field -> Text -> " <> tableNameToModelName tableName <> " -> " <> tableNameToModelName tableName <> "\n"
+        <> "assignField field value model = case field of \n" <> intercalate "\n" (map compileAssignField attributes) <> " \n"
+        <> "assign :: [(Field, Text)] -> " <> tableNameToModelName tableName <> " -> " <> tableNameToModelName tableName <> "\n"
+        <> "assign = undefined"
+    where
+        compileAssignField (Field name _) = "   " <>  tableNameToModelName name <> " -> model { " <> name <> " = value }"
 
+compileCombine table@(Table tableName attributes) =
+        "combine (" <> tableNameToModelName tableName <> " " <> (intercalate " " (attributesToArgs "arg" attributes)) <> ") (" <> tableNameToModelName tableName <> " " <> (intercalate " " (attributesToArgs "f" attributes)) <> ") = " <> tableNameToModelName tableName <> " " <> (intercalate " " (attributesToApplications attributes))
+    where
+        attributesToArgs :: Text -> [Attribute] -> [Text]
+        attributesToArgs prefix attributes = map (\n -> prefix <> tshow n) $ (map snd (zip attributes [0..]))
+        attributesToApplications attributes = map (\n -> "(f" <> tshow n <> " arg" <> tshow n <> ") ") $ (map snd (zip attributes [0..]))
 
 --compileAttributeBag :: Table -> Text
 --compileAttributeBag table@(Table name attributes) = "class To" <> tableNameToModelName name <> "Attributes where\n    to"
